Add SnackbarProvider screen reader and accessibility tests, cleanup
